// dont-eat-so-much-server/src/errors/types.ts
export enum ErrorType {
  INVALID_ACCESS_TOKEN = "INVALID_ACCESS_TOKEN",
  EXPIRED_ACCESS_TOKEN = "EXPIRED_ACCESS_TOKEN",
  INVALID_REFRESH_TOKEN = "INVALID_REFRESH_TOKEN",
  EXPIRED_REFRESH_TOKEN = "EXPIRED_REFRESH_TOKEN",
  INVALID_CREDENTIALS = "INVALID_CREDENTIALS",
  INVALID_EMAIL_FORMAT = "INVALID_EMAIL_FORMAT",
  USER_ALREADY_EXISTS = "USER_ALREADY_EXISTS",
  MEAL_NOT_FOUND = "MEAL_NOT_FOUND",
  INVALID_DATE_FORMAT = "INVALID_DATE_FORMAT",
  INVALID_BARCODE = "INVALID_BARCODE",
  BARCODE_ALREADY_EXISTS = "BARCODE_ALREADY_EXISTS",
}